requirements.txt
veracode-api-py>=0.9.5
logging-formatter-anticrlf>=1.2